use npx tsc

+++ packages/server/src/handlers/files.ts
@@ -1,7 +1,7 @@
+import { exec } from "child_process";
 import fs from "fs";
 import type { IncomingMessage, ServerResponse } from "http";
 import path from "path";
-import ts from "typescript";
 import { promisify } from "util";
 import { parseRequestBody, sendJSONResponse } from "../utils/request";
 
@@ -60,7 +60,7 @@ export const updateWorkflowFile = async (
 
   try {
     const resolvedFilePath = path.resolve(DEFAULT_WORKFLOW_FILE_PATH);
-    const result = validateTypeScriptFile(resolvedFilePath, content);
+    const result = await validateTypeScriptFile(resolvedFilePath, content);
 
     if (!result.isValid) {
       return sendJSONResponse(res, 400, {
@@ -77,7 +77,7 @@ export const updateWorkflowFile = async (
 };
 
 // use typescript package to validate the file syntax and imports
-function validateTypeScriptFile(filePath: string, content: string) {
+async function validateTypeScriptFile(filePath: string, content: string) {
   // Update workflow file directly will cause the server restart immediately.
   // So we create a temporary file with the same content in the same directory as the workflow file
   // This file will be used to validate the file syntax and imports. It will be deleted after validation.
@@ -87,34 +87,17 @@ function validateTypeScriptFile(filePath: string, content: string) {
   );
   fs.writeFileSync(tempFilePath, content);
 
-  // Create a TypeScript program to validate file by filePath
-  const program = ts.createProgram([tempFilePath], {
-    noEmit: true,
-    target: ts.ScriptTarget.ES2022,
-    module: ts.ModuleKind.ES2022,
-    moduleResolution: ts.ModuleResolutionKind.Bundler,
-    strict: true,
-    allowJs: false,
-    skipLibCheck: true,
-  });
-
-  // Get diagnostics (errors and warnings)
-  const diagnostics = ts.getPreEmitDiagnostics(program);
-
-  // delete the temp file
-  fs.unlinkSync(tempFilePath);
-
-  // Format and return errors
-  const errors = diagnostics.map((diagnostic) => {
-    const message = ts.flattenDiagnosticMessageText(
-      diagnostic.messageText,
-      "\n",
-    );
-    const { line, character } = diagnostic.file
-      ? diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start ?? 0)
-      : { line: 0, character: 0 };
-    return `Error at line ${line + 1}, character ${character + 1}: ${message}`;
-  });
+  const errors = [];
+  try {
+    const tscCommand = `npx tsc ${tempFilePath} --noEmit --skipLibCheck true`;
+    await promisify(exec)(tscCommand);
+  } catch (error) {
+    const errorMessage = (error as { stdout: string })?.stdout;
+    errors.push(errorMessage);
+  } finally {
+    // Clean up temporary file
+    if (fs.existsSync(tempFilePath)) fs.unlinkSync(tempFilePath);
+  }
 
   return {
     isValid: errors.length === 0,
